Fruity pebbles [1/4]
In the commit below we added the ability to change between Dark/Light and wallpaper
based themes.

https://github.com/DirtyUnicorns/android_frameworks_base/commit/d5de41e48b36951bbd96738b2cafd950efc5f068

Since then we've extended this to Settings, DU-Tweaks, Changelog and a few other
apps in system. We've added more styles and have gone all in on this theme thing.

One thing that was missing was the ability to change the system accent. While the
pixel accent was good enough, many users had asked us to add in the ability to
change the accents because they wanted something different. Several ROMs have gone
about this same thing in their own way but we didn't. We wanted to use what we already
had and just add on to it.

This commit mirrors what @maxwen did but with the accent, so thanks to him for that!

- Expose the DND panel 'close' drawable so we can theme it with the dark theme

Change-Id: I5bc383d749d34ba10c0cd96153d17e706b9dac3c