BatteryStats: Update WifiState even if EnergyInfo is invalid
CTS test case testWifiUpload and testWifiDownload fails as the
Wifi data and energy information statistics are not seen on
dumpsys batterystats --checkin.

If optional link layer statistics feature is not supported,
energyInfo returned is null. Even if energy information is
not available, tx and rx packet statistics needs to be
updated.

Fix the issue by allowing to dump data statistics even if
energy information data is not available.

Bug: 78197718
Test: test cts -m CtsIncidentHostTestCases with optional link layer statistics not supported
Change-Id: Ic516a47769e7c946fb7c5d54f34b9a1c143814c6
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Adesh15 <adesikha15@gmail.com>