Change the listening logic of MtpStorageManager
Need to adjust the current monitoring logic, in order to accurately
transfer file propertys to Windows. Current logic is that system
called handleAddedObject() immediately when the large file is created,
and transfer the defect state of the file, which will cause the Windows
system to get an incomplete data error when getting the file.

Test: boot phone,connect win10, enable mtp, push a file(>2M,such as ZIP)
Get this file from your phone to your computer，check the file.
Signed-off-by: katao <katao@xiaomi.com>

Change-Id: I6f5db78652d63ffb88535d49f5567f3c7f06c596
Signed-off-by: mydongistiny <jaysonedson@gmail.com>